Extract tables to component second step
