+++ src/components/tables/actions/create.js
@@ -1,29 +1,11 @@
-const Joi = require('joi');
-const Table = require('../Table');
+const Table = require('../model');
 const tablesSerializer = require('../tables-serializer');
-const JSONAPIDeserializer = require('jsonapi-serializer').Deserializer;
-
-const tableCreateSchema = Joi.object().keys({
-  seatsCount: Joi.number().min(1).max(30).required(),
-  name: Joi.string().required()
-});
 
 module.exports = async (ctx, next) => {
-  const deserializeOption = {
-    keyForAttribute: 'camelCase'
-  };
-  const payload = await new JSONAPIDeserializer(deserializeOption)
-    .deserialize(ctx.request.body);
-  const result = Joi.validate(payload, tableCreateSchema, { stripUnknown: true });
+  const newTable = await Table.create(ctx.request.body);
+  const response = tablesSerializer.serialize(newTable);
 
-  if (result.error) {
-    ctx.body = result.error;
-    ctx.status = 422;
-  } else {
-    const newTable = await Table.create(payload);
-    const response = tablesSerializer.serialize(newTable);
-    ctx.body = response
-  }
+  ctx.body = response;
 
   await next();
 }
