+++ src/components/tables/router.js
@@ -1,6 +1,9 @@
 const Router = require('koa-router');
 const bodyParser = require('koa-bodyparser');
-const { create, findAll } = require('./actions')
+const JSONApiParser = require('./middlewares/JSONApiParser');
+const { create, findAll } = require('./actions');
+const { create: createValidator } = require('./validators');
+const validate = require('./middlewares/validate');
 
 const router = new Router();
 
@@ -10,6 +13,8 @@ router.get('/api/v1/tables',
 
 router.post('/api/v1/tables',
   bodyParser(),
+  JSONApiParser,
+  validate(createValidator),
   create
 );
